requirements.txt
fire>=0.1.3
regex==2017.4.5
requests==2.21.0
tqdm==4.31.1


tensorflow-gpu>=2.0
scikit-learn>=0.22.0
scipy>=1.3.2
pandas>=1.0.1
pandas-datareader>=0.8.0
matplotlib==3.2.1
pillow>=6.2.2
h5py>=1.8.4
pyyaml==5.3
flask>=1.1.1
boto3>= 1.12.24
        
